+++ src/covariate_model.cpp
@@ -12,10 +12,10 @@ double log_lik(NumericMatrix humans, NumericMatrix phi, NumericVector p) {
   double scaleP = 1 / std::accumulate(P.begin(), P.end(), 0.0);
 
   // run through the rows of the human matrix
-  for (size_t h = 0; h < humans.nrow(); ++h) {
+  for (int h = 0; h < humans.nrow(); ++h) {
     // calculate the likelihood for this human isolate
     double lik_h = 0;
-    for (size_t j = 0; j < P.size(); j++) {
+    for (int j = 0; j < P.size(); j++) {
       lik_h += P[j] * phi(humans(h,0)-1, j);
     }
     loglik += humans(h,1) * log(lik_h * scaleP);
